tools/requirements.txt
PyMuPDF==1.22.5
Pillow==10.1.0
watchdog==3.0.0
